pyproject.toml
[tool.poetry]
name = "codeleap-code-test"
version = "0.1.0"
description = "A basic blogging API"
authors = ["Felipe Matheus <contato@flipeip.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.4"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
